backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==2.3.3
openpyxl==3.1.5
google-generativeai==0.8.5
python-dotenv==1.2.1
pymongo==4.15.4
dnspython==2.8.0
